requirements.txt
clickclick>=0.9
keyring
PyYAML
requests
